% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_trie}
\alias{create_trie}
\title{Create a Trie}
\usage{
create_trie(keys, values)
}
\arguments{
\item{keys}{a character vector containing the keys to match against}

\item{values}{an atomic vector of any type, containing the values to pair with
\code{keys}. Must be the same length as \code{keys}.}
}
\value{
a `trie` object.
}
\description{
\code{create_trie} creates a trie (a key-value store optimised
for matching) out of a provided character vector of keys, and a numeric,
character, logical or integer vector of values (both the same length).
}

