% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getters}
\alias{get_keys}
\alias{get_keys.bool_trie}
\alias{get_keys.integer_trie}
\alias{get_keys.numeric_trie}
\alias{get_keys.string_trie}
\alias{get_values}
\alias{get_values.bool_trie}
\alias{get_values.integer_trie}
\alias{get_values.numeric_trie}
\alias{get_values.string_trie}
\alias{getters}
\title{Trie Getters}
\usage{
get_keys(trie)

get_values(trie)

\method{get_keys}{numeric_trie}(trie)

\method{get_keys}{integer_trie}(trie)

\method{get_keys}{string_trie}(trie)

\method{get_keys}{bool_trie}(trie)

\method{get_values}{numeric_trie}(trie)

\method{get_values}{integer_trie}(trie)

\method{get_values}{string_trie}(trie)

\method{get_values}{bool_trie}(trie)
}
\arguments{
\item{trie}{A trie object.}
}
\value{
An atomic vector of keys or values stored in the trie.
}
\description{
"Getters" for the data stored in a trie object. \code{get_keys}
gets the keys, \code{get_values} gets the values.
}

