% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triebeard.R
\docType{package}
\name{triebeard}
\alias{triebeard}
\alias{triebeard-package}
\title{Radix trees in Rcpp}
\description{
This package provides access to Radix tree (or "trie") structures in Rcpp. At
a later date it will hopefully provide them in R, too.
}

