% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getters}
\alias{get_keys}
\alias{get_values}
\alias{getters}
\title{Trie Getters}
\usage{
get_keys(trie)

get_values(trie)
}
\arguments{
\item{trie}{A trie object, created with \code{\link{trie}}.}
}
\value{
An atomic vector of keys or values stored in the trie.
}
\description{
"Getters" for the data stored in a trie object. \code{get_keys}
gets the keys, \code{get_values} gets the values.
}

