% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.R
\name{longest_match}
\alias{longest_match}
\title{Find the longest match in a trie}
\usage{
longest_match(trie, to_match)
}
\description{
boop
}
\examples{
trie <- create_trie(keys = c("afford", "affair", "available", "binary", "bind", "blind"),
                   values = c(1, 2, 3, 4, 5, 6))
longest_match(trie, "binder")
rm(trie)
}

